requirements.txt
Django==3.2.25
psycopg2_binary==2.9.9
Pillow==9.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
django-multiselectfield==0.1.12
django-import-export==2.8.0
django-cors-headers==3.10.0 
django-cors-middleware==1.5.0
drf_yasg==1.21.7
